add Export functionality
When result_path is set, builders must copy their output from the
build_root to a directory that will be shared over nfs.  For now this
is accomplished with `cp -a`.  It may be necessary to select files
with globs in the future.

PKG_CONFIG settings were added to AutoBuilder, and were tested against
an m32 build.

+++ scripts/build_local.py
@@ -51,8 +51,10 @@ def main():
     vdict = vars(args)
     del vdict["project"]
     o.__dict__.update(vdict)
+    sys.argv = ["bogus"] + o.to_string().split()
 
-    bs.rmtree(bs.ProjectMap().build_root())
+    if "clean" in o.action:
+        bs.rmtree(bs.ProjectMap().build_root())
 
     graph = bs.DependencyGraph(project, o)
     ready = graph.ready_builds()
